Query Additions
Added the with-query function that acts exactly like the with-session
method and allows controlled closing of the query object when it is not
needed anymore. Problem is though, this kind makes code a bit nexted
when you use both macros directly (with-session [] (with-query [])) so
there must be a way to bring them together and make the code look a bit
nicer.

Added the results method as well.  Which attempts to make a lazy-ish
sequence from the result set so we aren't forced down the next?/more
route like a half baked Java iterator.  I say lazy-ish because the
actual implementation in the Java client already loads everything into
an ArrayList cache and iterates over that.  Perhaps in future we can
manage this ourselves directly