Renaming modules and code refactoring
In order to maintain consistency across the existing BaseX clients and
server vocabulary client and notification modules are now simply the
session module.  So we have session specific features and query specific
features.

The has been updated in the tests and the associated README sections

Also reduced the amoutn of code in the results method of query. Rather
than replying on the more? method of the underlying Query class it just
lazily generates the next value in the query cache until it becomes
nil.